* Allow for mime-type exporting with named exporters (eg. say I want image/png from   THAT special exporter) with test "Alien Glow" exporter * Allow for exporting to special "application/x-qt-image" for native image formats * discovered bug for de-installing add-ons: will not deinstall the respective plugin(s)   correctly (still needs fix, duplicate because of arches....etc) * Small fixes (among others klfconfig.Core.thisVersion*FirstRun) * more small fixes
git-svn-id: https://svn.code.sf.net/p/klatexformula/code/trunk/klatexformula@496 4112663c-bd99-11de-a44c-fbaffaffe1b7

+++ src/klfbackend/klfbackend.h
@@ -144,14 +144,22 @@ public:
      * be reported as real errors. */
     QString epstopdfexec;
 
-    /** The number of postscript points to add to top side of the resulting EPS boundingbox */
-    int tborderoffset;
-    /** The number of postscript points to add to right side of the resulting EPS boundingbox */
-    int rborderoffset;
-    /** The number of postscript points to add to bottom side of the resulting EPS boundingbox */
-    int bborderoffset;
-    /** The number of postscript points to add to left side of the resulting EPS boundingbox */
-    int lborderoffset;
+    /** The number of postscript points to add to top side of the resulting EPS boundingbox.
+     * \note Currently this value is rounded off to the nearest integer. The type \c double is
+     *   just an anticipation of hi-resolution bounding box adjustment. */
+    double tborderoffset;
+    /** The number of postscript points to add to right side of the resulting EPS boundingbox
+     * \note Currently this value is rounded off to the nearest integer. The type \c double is
+     *   just an anticipation of hi-resolution bounding box adjustment. */
+    double rborderoffset;
+    /** The number of postscript points to add to bottom side of the resulting EPS boundingbox
+     * \note Currently this value is rounded off to the nearest integer. The type \c double is
+     *   just an anticipation of hi-resolution bounding box adjustment. */
+    double bborderoffset;
+    /** The number of postscript points to add to left side of the resulting EPS boundingbox
+     * \note Currently this value is rounded off to the nearest integer. The type \c double is
+     *   just an anticipation of hi-resolution bounding box adjustment. */
+    double lborderoffset;
 
     /** Strip away fonts in favor of vectorially outlining them with gs.
      *
